feat: #34 add separate Image lookup logic for MissionSubmit details
