+++ src/main/java/laughcandidate/yellowribbonbe/admin/dto/response/MissionSubmitResponse.java
@@ -17,15 +17,15 @@ public record MissionSubmitResponse(
         String imageUuid,
         LocalDateTime submittedAt
 ) {
-    public static MissionSubmitResponse from(MissionSubmit missionSubmit) {
+    public static MissionSubmitResponse from(MissionSubmit missionSubmit, Long imageId, String imageUuid) {
         return MissionSubmitResponse.builder()
                 .missionSubmitId(missionSubmit.getId())
                 .status(missionSubmit.getStatus())
                 .reason(missionSubmit.getReason())
                 .missionCategory(missionSubmit.getMission().getCategory())
                 .missionDescription(missionSubmit.getMission().getDescription())
-                .imageId(missionSubmit.getImage().getId())
-                .imageUuid(missionSubmit.getImage().getUuid())
+                .imageId(imageId)
+                .imageUuid(imageUuid)
                 .submittedAt(missionSubmit.getCreatedAt())
                 .build();
     }
