+++ src/main/java/laughcandidate/yellowribbonbe/mission/repository/MissionSubmitRepository.java
@@ -3,6 +3,7 @@ package laughcandidate.yellowribbonbe.mission.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import laughcandidate.yellowribbonbe.mission.entity.MissionSubmit;
+import laughcandidate.yellowribbonbe.mission.repository.custom.MissionCustomRepository;
 
-public interface MissionSubmitRepository extends JpaRepository<MissionSubmit, Long>, MissionSubmitRepositoryCustom {
+public interface MissionSubmitRepository extends JpaRepository<MissionSubmit, Long>, MissionCustomRepository {
 }
